moving geocoder to docker, adding es geocoder to python rather than importing node script
