+++ docker/geocoder/.gitignore
