+++ docker/geocoder/Dockerfile
@@ -1,0 +1,2 @@
+FROM python:3-onbuild
+CMD ["python", "./daemon.py"]
